Fix min function type error
Corrected a compilation error in `handleScan()` where `std::min()` was called with arguments of conflicting types (`int16_t` and `int`). Explicitly cast `n` to `int` to resolve the type mismatch and allow compilation.

+++ Firmware_ESP32_Sensores_v4.3.3_WEB_IMPROVED.ino
@@ -1074,7 +1074,7 @@ void handleScan() {
     logMessage(LOG_INFO, "✅ Scan completo: " + String(n) + " redes");
     
     // Limitar a 15 redes mais fortes
-    int maxNetworks = min(n, 15);
+    int maxNetworks = min((int)n, 15);
     
     // Construir JSON
     String json = "{\"status\":\"complete\",\"networks\":[";
